+++ src/main/java/AdamException.java
@@ -1,0 +1,13 @@
+public class AdamException extends RuntimeException{
+    public AdamException(){}
+    public AdamException(String message){
+        super(message);
+    }
+    public AdamException(Throwable cause){
+        super(cause);
+    }
+
+    public AdamException(String message, Throwable cause){
+        super(message,cause);
+    }
+}
